+++ CHANGES.rst
@@ -10,6 +10,8 @@ Version 2.3.3
 Improvements
 ^^^^^^^^^^^^
 
+- Add a new "Until approved" option for a registration form's "Modification allowed"
+  setting (:pr:`4740`, thanks :user:`vasantvohra`)
 - Show last login time in dashboard (:pr:`4735`, thanks :user:`vasantvohra`)
 - Allow Markdown in the "Message for complete registrations" option of a registration
   form (:pr:`4741`)